3/2/requirements.txt
apache-airflow-providers-apache-cassandra==3.7.0
cassandra-driver==3.29.2
